Fixed kernel when --debug and no stopwatch
- Using the regular event dispatcher
