+++ .formatter.yml
@@ -7,7 +7,7 @@ use-sort:
 strict: true
 header: |
     /*
-     * This file is part of the Symfony Async Kernel
+     * This file is part of the Drift Http Kernel
      *
      * For the full copyright and license information, please view the LICENSE
      * file that was distributed with this source code.
