+++ .php_cs
@@ -13,6 +13,7 @@ return PhpCsFixer\Config::create()
         '@PSR2' => true,
         '@Symfony' => true,
         'single_line_after_imports' => false,
+        'no_superfluous_phpdoc_tags' => false
     ])
     ->setFinder($finder)
 ;
