+++ AsyncKernel.php
@@ -1,7 +1,7 @@
 <?php
 
 /*
- * This file is part of the Symfony Async Kernel
+ * This file is part of the Drift Http Kernel
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -117,6 +117,8 @@ abstract class AsyncKernel extends Kernel implements CompilerPassInterface
     private function processEventDispatcherDebug(ContainerBuilder $container)
     {
         if (!$container->hasDefinition('debug.event_dispatcher')) {
+            $this->processEventDispatcher($container);
+
             return;
         }
