+++ Tests/Base/AutowiringDebugTest.php
@@ -1,12 +1,25 @@
 <?php
 
+/*
+ * This file is part of the Drift Http Kernel
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * Feel free to edit as you please, and have fun.
+ *
+ * @author Marc Morera <yuhu@mmoreram.com>
+ */
 
-namespace Drift\HttpKernel\Tests;
+declare(strict_types=1);
 
+namespace Drift\HttpKernel\Tests\Base;
+
+use Drift\HttpKernel\Tests\AsyncKernelFunctionalTest;
 use Drift\HttpKernel\Tests\Services\AService;
 
 /**
- * Class AutowiringDebugTest
+ * Class AutowiringDebugTest.
  */
 class AutowiringDebugTest extends AsyncKernelFunctionalTest
 {
@@ -21,24 +34,24 @@ class AutowiringDebugTest extends AsyncKernelFunctionalTest
     {
         $configuration = parent::decorateConfiguration($configuration);
         $configuration['imports'] = [
-            ['resource' => dirname(__FILE__) . '/autowiring.yml']
+            ['resource' => dirname(__FILE__).'/../autowiring.yml'],
         ];
 
         return $configuration;
     }
 
     /**
-     * Test autowiring
+     * Test autowiring.
      */
     public function testAutowiring()
     {
         $aService = $this->get(AService::class);
         $this->assertTrue($aService->equal);
-        $this->assertTrue($aService->isTraceable);
+        $this->assertFalse($aService->isTraceable);
     }
 
     /**
-     * Kernel in debug mode
+     * Kernel in debug mode.
      *
      * @return bool
      */
