+++ Tests/Services/AService.php
@@ -1,5 +1,17 @@
 <?php
 
+/*
+ * This file is part of the Drift Http Kernel
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * Feel free to edit as you please, and have fun.
+ *
+ * @author Marc Morera <yuhu@mmoreram.com>
+ */
+
+declare(strict_types=1);
 
 namespace Drift\HttpKernel\Tests\Services;
 
@@ -10,7 +22,7 @@ use React\Filesystem\Filesystem;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 
 /**
- * Class AService
+ * Class AService.
  */
 final class AService
 {
@@ -20,18 +32,17 @@ final class AService
     /**
      * AService constructor.
      *
-     * @param EventDispatcherInterface $dispatcher
+     * @param EventDispatcherInterface      $dispatcher
      * @param AsyncEventDispatcherInterface $dispatcher
-     * @param LoopInterface $loop
-     * @param Filesystem $filesystem
+     * @param LoopInterface                 $loop
+     * @param Filesystem                    $filesystem
      */
     public function __construct(
         EventDispatcherInterface $dispatcher1,
         AsyncEventDispatcherInterface $dispatcher2,
         LoopInterface $loop,
         Filesystem $filesystem
-    )
-    {
+    ) {
         $this->equal = ($dispatcher1 === $dispatcher2);
         $this->isTraceable = ($dispatcher2 instanceof TraceableAsyncEventDispatcher);
     }
