+++ composer.json
@@ -12,18 +12,18 @@
     "require": {
         "php": "^7.2.9",
 
-        "symfony/http-kernel": "^4.3 || ^5.0",
-        "symfony/dependency-injection": "^4.3 || ^5.0",
-        "react/promise": "*",
-        "react/filesystem": "*"
+        "symfony/http-kernel": "^4.4 || ^5.0",
+        "symfony/dependency-injection": "^4.4 || ^5.0",
+        "react/promise": "^2.7",
+        "react/filesystem": "^0.1",
+        "phpunit/phpunit": "7.5.17"
     },
     "require-dev": {
         "mmoreram/symfony-bundle-dependencies": "^2.2.0",
         "mmoreram/base-bundle": "^2.0.0",
-        "symfony/yaml": "^5.0",
-        "symfony/console": "^5.0",
-        "symfony/stopwatch": "^5.0",
-        "clue/block-react": "*"
+        "symfony/yaml": "^4.4 || ^5.0",
+        "symfony/console": "^4.4 || ^5.0",
+        "clue/block-react": "^1.3"
     },
     "autoload": {
         "psr-4": {
@@ -32,6 +32,5 @@
         "exclude-from-classmap": [
             "/Tests/"
         ]
-    },
-    "minimum-stability": "dev"
+    }
 }
